Make ES2 support dynamic
Instead of the rigid QT_OPENGL_ES_2 check, which is not suitable
for -opengl dynamic builds on Windows at all, make ES2 support
available always and use it whenever the context is an ES context.

The usage of QOpenGLFunctions_ES2 must be avoided since it is tied
to ES-only builds. In practice QOpenGLFunctions serves the same
purpose since Qt 5.4, so switch to that.

This also enables easy testing of the desktop and ES rendering path
side by side on Windows: a -opengl dynamic build, setting QT_OPENGL=angle
and a default 2.0 QOpenGLContext is all that is needed to switch over to ES2.

The check for the presence of an ES VAO extension is removed. Some
implementations will just not have it so we cannot count on it.

Change-Id: I57239f73375520fe186a4c7e5f991bc32ceea0ee
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>