Fix viewport when using an external FBO
When using Scene3D (i.e. QQuickFramebufferObject), we are rendering to
a non-default FBO but will not have a corresponding render target according
to our internal bookkeeping. Avoid glViewport(0, 0, -1, -1) by using
the surface size in this case.

While it might be tempting to add code for querying the framebuffer (attachment)
size dynamically instead, that is impossible with ES2 since the size of textures
cannot be queried there.

Change-Id: I8b2e39a8a23e5463f359539d794dfe15b0434abc
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>