Clean up GL2-ES2 shaders
Unify light intensity in the phong material and shaders. The GL3 version had
1, 1, 1 whereas the other two used 1, 0, 1. The former is more sensible, so
standardize on that. We also remove the default uniform values for GL2 and 3.
The defaults are now coming from one central place in PhongMaterial.qml.

All fragment shaders need to be revamped since these are common for OpenGL 2.0
and OpenGL ES 2.0. The only way to get something that compiles with both is to
have the float/vecX precision qualifiers inline. This is in line with how
ShaderEffect items work with Qt Quick. On non-ES Qt will dynamically add the
necessary empty defines for highp etc.

Remove also comments about uniform blocks (since it is not an option for ES2) and
clean up the default values (which are coming from QML).

Change-Id: If9862b4f2faf0b877da4032dcfcd123f259aa168
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>