QAbstractTextureImage + QAbstractTextureDataFunctor
A QAbstractTextureProvider will be completed by a list of QTextureImage.
Each QTextureImage provides a layer, mipmapLevel, cubeFace and source
properties.
Each of the QTextureImage will also provide a QAbstractTextureDataFunctor that
will be used to load the actual texture data in an Aspect job as well as
compared to other functors for texture sharing.

Change-Id: Id17fd83da78b906d71320b629fc70f6aedb6a63e
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>