RenderTexture: update list of texture image id to texture image handles
That required a custom backend functor creator. QRenderAspect was updated to
use the correct functor.

This will provide faster lookups. Also introduced TextureDNA.
The goal is to build a unique key based on the texture properties and each
texture image (including layer position, mipmaplevel and cubemapface).
Then when applying a texture it will just be a matter of comparing hashes.

Change-Id: Ic07aea6bc1cbcb1059ec5bcdd458ccc51e1feda9
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>