Remove default uniform values for lightPosition and intensity
The ES2 variants of the shaders have a default uniform value of 0.
To be compatible with the other versions, start specifying the
values for lightPosition and Intensity from C++, just like we already
do from QML. This way the default values are coming from one single
unified place regardless of the shader version in use, and examples
like simple-cpp will work identically to simple-qml, even on ES2.

Change-Id: If60185d0d3039e4354da8cc6e6b21a0db0bae963
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>