Avoid issue draw calls when the buffer is not yet ready
Follow-up to the previous VAO fix. Issuing the draw call always is wrong
and even crashes some embedded devices' GLES implementation because the
buffer and attributes may not yet be set up due to the condition above
(when m_parameterAttributeToShaderNames is still empty).

Make it conditional by drawing only if we did go through the buffer
setup or the VAO is supported and available.

This makes the renderer functional on plain GLES 2.0 implementations
that are often found in embedded devices.

Change-Id: Ib550c2321fe05c75c998ce2f38079bdd703b03d3
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>