Fix image format for textures
We always convert to RGBA8888 so the only acceptable combination is GL_RGBA - GL_RGBA8.

Previously we were checking the presence of the alpha channel in the original image. This
is quite wrong since we end up with GL_RGBA - GL_RGB8 type of combinations in case the
input has no alpha. That is not a valid combination.

This makes images without an alpha channel, e.g. image formats that provide Format_RGB32
QImages, working.

Change-Id: I2631071134c5224c52b64cff2f00ef6ae166762b
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>