MouseInput API and example mockup
The meaning of exclusiveFocus is not set in stone for now. We know some kind
of exclusive focus will be needed but defining it precisely will be done as
the input aspect takes shape.

QMouseController -> dispatch events to QMouseInput, specify the camera to be
used to perform the picks.
QMouseInput -> Component that offers a way to react to mouse events
Q3DMouseEvent/Q3DWheelEvent -> Qt3D mapping of QMouseEvent/QWheelEvent

Change-Id: I462bea2932d4bad6595d04a75ea1e0a67c7c8ba0
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>