Do not build render jobs without a surface
Examples like wireframe tend to crash on exit. This is because the platform
surface filter sets the renderer's surface back to null, while we are trying
to build some jobs still. Even though everything gets shutdown soon after that,
we have to somehow prevent any operations relying on the surface in the meantime.

This is also an opportunity to avoid exposing the surface pointer from Renderer.
Instead, communicate the surface size to the render views.

Change-Id: Ifc739ec50e3f6a67d20bfd0dc1913a12b906a8ab
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>