Fix default OpenGL filter settings
The API should be ES in an ES-only build of Qt, or when dynamically choosing
ANGLE on Windows.

The profile should be NoProfile. This is particularly important for systems where
requesting a version like 2.0 gives a < 3.2 no profile context instead of >= 3.2
compatibility. The default filter should match both, hence using Compatibility is
not an option.

Change-Id: Ie8f20b0f09d8d6f6569cfda0f8c2b7217a2e23df
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>