Fix modifier key handling in input aspect
Relying on Alt press/releases has the annoying side effect that some systems (X11)
will not give a KeyRelease for Alt when pressing Alt+Tab.

It is better to handle such keys based on the modifier state. This avoids getting into
an "Alt is stuck" situation when switching back to the application's window.

Change-Id: I8e1675d9d4093e162fe0e42a2def26737d7a7467
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>