Avoid crashing when becoming non-exposed
On Android Qt3D apps used to crash when they got moved to background.
beginDrawing() correctly recognized that the window was not exposed anymore, but
some of the cleanup for the render queue was not done in this case.

This is now corrected, so apps will survive. It is not the full solution because we
still continue to attempt to render which is a waste of CPU time. Proper pause and resume
will have to be implemented later, similarly to the Quick scenegraph render loops, based
on the expose events.

Change-Id: Ia567d891fc746bb6851524e0f8a998a8fd53f8b1
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>