Remove copyright headers from the ES3-specific shadow map shaders
Not having the headers is fine since the built-in shaders do not have
them either, neither in Quick nor in Qt3D.

The main issue with ES is that in ESSL3 the spec states that the version
directive is only allowed in the first line and no comments are allowed before.
This is not true for standard GLSL where the spec states the comments and
whitespace are allowed before.

Mesa does not seem to enforce this, but we have already found at least one
GL ES 3.0 implementation which does.

Change-Id: I6ac3f949eeb44a005f69cc03f7d6f6eb5b577f2e
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>