Link assimp scene plugin on hosts with a system Qt5 installed
Using PKGCONFIG_PRIVATE results in the -L/usr/lib -lassimp being placed
at the end of the linker command line rather than the start as with
PKGCONFIG.

The latter breaks the build if there is an installed system Qt5 of an
older version in the same directory as found by the assimp pkgconfig
file. The breakage is caused by undefined symbols in the older version
of Qt in the system directory.

Change-Id: I691b437eadc8b6fec385374d0910344ed194933b
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>