Remove unneeded signal
This was also triggering a warning about conflicting signal names on the
unextended QTransform and the extension Quick3DTransform types. Only
seen in developer builds as a result of the connect call removed in
this commit.

None of the other QQmlListProperty properties have notify signals so
remove this one too to avoid the conflict.

Change-Id: I877a83e5feec89f97fced0e0b6c024e882555846
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>