Improve bigmodel example by using a QtQuick ListModel
When using a number as a model, when the model changes,
a new model is recreated which induces all delegates to
be cleared and recreated.
Using a ListModel however, the model reference remains
the same, only its content is updated which implies that
already created delegates are kept which is much more
efficient.

This example could further be improved by making the
delegate more lightweight by defining components in the
scene root and referencing those as components in the
delegate. However this would complexify this example
which is there rather as a how to than a performance
oriented one.

Change-Id: I542889a25b65f5de4579288b00b56c9c52667764
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>