Support iOS FBO handling
iOS does not have a default FBO. One gets created by the QPA plugin.
But Qt3D cannot make assumptions about what the id of the default.
So QGraphicsContext tracks the default FBO id. That is used in viewport
size handling and render target activation.
It is also used in the renderer to decide is swap buffers is needed.

Change-Id: Icb3c44fde2a6c5220550f2745caabe49ec955cf1
Reviewed-by: James Turner <james.turner@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>