Properly shutdown the renderer
Now all pure Qt3D examples have a 0 exit code.

Still need some work for QtQuick/Scene3D examples. The problem being that we
hit the QOpenGLDebugger issue since we are unable to shutdown properly the
Qt3D renderer because Scene3D being a QQuickItem is destroyed before the
QQuickWindow and we never have a chance to receive the sceneGraphInvalidated
signal.

Change-Id: I1d30e0d883057630a55cd3f99eaa6bdb65a8df26
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>