Add support for loading scenes from glTF 0.8 files
This patch includes a new scene parser plugin which enables the loading
of glTF files (using the 0.8 spec) into Qt3D scenes.  The gltf example
has been updated to work with the current APIs and the asset files it
used have been updated to match the glTF 0.8 spec requrements.

Previously an unused copy of the GLTF scene parser was located in
render/io.  It was being built, but not used as only plugin based
sceneloaders were loaded by the renderer.  Now it should be usable and
always available.

Change-Id: Ic2e31e2b63a871559aad9bad90ec2820988a1571
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>