Rework Renderer to use BufferFunctor
Each buffer when it receives a new functor adds itself with the buffer
manager. This then allows to build LoadBufferJob instances for each buffer to
be loaded. The buffers can then be loaded before a frame is rendered.

The GeometryRendererManager has been moved back with other managers and the
BufferManager takes its place to handle the additional management of dirty
buffers.

Renderer updated accordingly.

Change-Id: I5df839b56d18de077e1ff2921cb481ed8f613567
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>