glTF: Support generating and using Core profile and GL2 shaders
This change effects both the qgltf tool and the gltfparser.

For the qgltf tool I change the way core profile shaders are defined.
Rather than adding an additional property to the shader object, I create
the necessary shaders, programs, and techniques, as well as add an
additional two properties to the Material so that that multiple
techniques can be used.  By default the shaders support OpenGL ES2 and
is compatible with the glTF 0.8 standard.

For the gltfparser, I now create additional QTechniques from the two new
properties attached to the material object when they exist.  This allows
for the use of OpenGLFilters to choose which technique to use under
certain situations.  So now it is possible to render modules processed
with qgltf in the OpenGL Core profile senario.

Change-Id: Iae10f93a03c29a7bb6fe0a3dfd0d568ba0043f12
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>