Fix Attribute Parameter mapping
When using a shader that has different attribute names than the expected
standard attributes, and there is a mapping between the two, currently
the attributes fail to be found because we are still using the standard
name instead of the mapped one.  This change fixes this behavior to allow
attribute names to be mapped as intended.

Change-Id: I8f264e348dab5257c55206065bc2900c8c3cff46
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>