Add new collision aspect
This will implement a simple collision-only (no physics) aspect that
can be used to detect when collisions occur between entities that
aggregate collider components.

It is also intended to provide a concrete implementation of the ray cast
service to permit users to perform ray casts and to also allow the input
aspect to correctly deliver input events to entities.

Change-Id: I20012ebdb0b3f1b2014ccdab039e87973d112c16
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>