Always Flip UV Coordinates in Objloader
Currently if you use a QMesh that has its geometry loaded from an obj
file, the UV coordinates are such that the default loading behavior for
textures is inversed over the Y axis, and now any textures provide must
be preprocessed to be inversed over the Y axis as well.  This is not
consistent with the behavior of our scene loaders which allow for
Textures to be loaded normally (right-side-up) and still get sampled
correctly.  This is because the scene loaders make sure to flip the UV
coordinates when loading the asset.  To be consistent, we now the UV
coordinates when loading QMesh geometry with the Objloader.

[ChangeLog][Qt3DRenderer][Objloader] UV Coordinates are now always
flipped when loading QMesh Geometry from an obj source.

Task-number: QTBUG-46815
Change-Id: I8c3ee5de5c8bbaf5967ef81059ed97eb85901ac2
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>