QNode: Introduce enabled property
Previously there was only and enable property for the QNode subclass
QComponent, but now all QNode subclasses should also have an enable
property. This means that it is now possible to disable a QEntity which
would also disable all of its children.  This is a bit tricky though
because each Aspect needs to now check if the corresponding QEntities
are disabled or not before processing.  This has been accounted for
with RenderEntity in the Renderer already, but will need to be respected
in each Aspect.

This behavior is similar to the Unity3D GameObject active property which
disabled the entity and disables all components.

Change-Id: I5233b2b43f04aa5dfa7910eac6a88a9da656a997
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>