Remove VAO assert from GL2 helper
GraphicsContext contains appropriate checks for VAO support. Asserting
here is not necessary, especially considering that the ARB extension
is never present on OS X (with older machines at least).

This assert should have been removed together with the introduction
of VAO-less rendering support for ES2.

This allows running Qt 3D apps with a default 2.0 context on OS X
in case only the APPLE extension is available.

Change-Id: If3dbf4548de65751bc214706e9569e2b7b42e326
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>