Manage additional attributes separately from base ones
This is needed because hen the QML engine first creates a
QCylinderGeometry and applies a binding such as

CylinderGeometry {
   attributes: [ myAttribute ]
}

it first calls the clear function on the QQmlListProperty which removes
all of the attributes for the base cylinder geometry before it calls
the append function to add the myAttribute. This is not what we want so
we explicitly manage the extra attributes added via the list property.

This differs from the case we are more used to in QML where setting a
list property of a contained object from an outer object does not call
the clear function first but rather just does the append.

Change-Id: I4e858f02c0114d667a325735788cb711ad8b8b37
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>