Fix: Multi Viewport with Scene3D
When using an external FBO (Scene3D with QtQuick) we may still want to use
multiple subviewports. In this case, we need to retrieve the size of the FBO's
first attachments (either a RenderBuffer in case of multisampled FBO or a
Texture2D) to then properly call glViewport.

Note: This is partially supported by GL2 and ES2.

Change-Id: Ie705e093c8edbcb08010d45086be9ad3be741629
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>