Scene3D/Renderer fix
Only allow a proceedToNextFrame call when we are sure a
frame has been previously submitted (Scene3D) case.
Previously, in the Scene3D case, if the RenderViewJob
took too long to complete, several calls to proceedToNextFrame
could be made in the meantine which eventually resulted in the
RenderQueue being reset, memory being released and corrupting the
data produced by the still running RenderViewJob.

Change-Id: I28b343d673305f7a883014922e5d78d5e730f630
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>