Resurrect bounding volumes
For BackToFront sorting to work we need to calculate a depth value based
on the entity's bounding spheres. This is currently missing as there is no
AABB from which the RenderEntity's localBoundingVolume could be
calculated.
Bring it back.
RenderEntity contains two world bounding volumes from now on: just
having one that includes the children is not ideal because in some cases
we want to know the entitys own bounding volume not including the children.
This is the case when building render commands for example.
Change-Id: I20f5bccdc3f868bcbd5faa3164f0e9995a3171ec
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>