Add first person camera controller mode and make it the default
For most scenes this is what people want in preference to the free-look
mode. Keeping a consistent up vector is more natural for most people.

This is a temporary hack until we get the InputManager completed and
we have a proper abstraction for controllers. After all, who says the
inputs should control the camera. They should be able to control any
Entity with a Transform. This also requires the Transform refactoring
to be completed before we can tie both sides together.

Change-Id: If8f02c66febcdee31c13ee46b6651a0e923e8d09
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>