Doc: Move the doc artifacts outside of the submodule directory
The qdocconf now lives in 'src/doc' because the documentation
project defined by it covers the whole Qt 3D module. This change
enables qmake to use 'qt3d' as the output directory for doc make
targets. The new 'doc' directory also includes the relevant images,
snippets, and the qdoc files.

Change-Id: Ibc6d0912f8b1e665767e97630b3e5509e10c5710
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>