assimp: Fix building with mingw older than 4.9
On Windows when building with mingw versions older than 4.9 the assembly
of IFCReaderGen.o fails because it reaches the section limit.  Normally
we would use the -mbig-obj flag to the assembler, but older versions of
GNU as don't have support for this.  So to work around this
IFCReaderGen.cpp is split into two separate files, which keeps the
sections below the limit for each object file.

Task-number: QTBUG-48902
Change-Id: If0cae29c10a50e3a5edf85fe21b8d2528971eb83
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>