qgltf: Do not use system-zlib for host tools
It was wrong of us to assume that a QT_CONFIG containing system-zlib
meant that we could use zlib for a host_build.  There are cases where
the host does not have zlib, but the target does (cross compilation). It
also breaks the rules for host_build tools.  Fortunately the bootstrap
library does provide the zlib symbols, so we just need to make sure we
include that.

Change-Id: Ibe25a22676c47909d95a83df46e311bd486fc9f7
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>