Remove incomplete collision aspect
It does not make sense to release the collision aspect as is as the API
is not final, and no backends have been implemented yet.  The collision
aspect is also out of scope for the current release, and for now we will
just support rudimentary collision support to enable picking by casting
rays at bounding spheres in the render aspect.

Task-number: QTBUG-49091
Change-Id: I6aa9e2a38d5e4974ea22b87aead1c0af93585f39
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>