Setting matrix property to decompose to S, R, T
Updating the matrix property now decomposes the matrix to
calculate the scale, rotation and translation properties that
get sent to the backend.

The test example will be updated to show how to build a matrix
from a series of "sub transformations" in an imperative way. This
will reinforce in the API the fact that we can't decompose an
affine transform into an arbitrary set of sub transforms. That is
the API will emphasise the one way nature.

The decomposition functions were provided by Konstantin Ritt.

Note: mouse picking unit tests skipped for now

Change-Id: Ibc259dacac7a3dc32f4b03b375607291d998601c
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>