Renderer: fix empty VAO case
Under some cases a VAO could be allocated but never created as at a given
point in time there was no attributes found for the current shader (which can
happen if the shader isn't yet loaded). Then the next frame, the attributes
would have been loaded and the VAO allocated and the renderer would assume
then that the VAO had been created. Eventually that would then crash as a draw
call would be performed with an empty and never created VAO.

Now check that a VAO is created and if not request creation to make sure that
the VAO is always created when setting the attributes/trying to draw.

Change-Id: Idf3fa45fabc574c0a0e82a38283c5848b32e15a4
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>