Extend qml value type for QMatrix4x4 to be more useful
Add functions for applying common transformations: translate, scale,
rotate and lookAt that are needed when building affine transformation
matrices.

It would be very useful to update the QtObject::method_matrix4x4()
function in qtdeclarative to allow calling Qt.matrix4x4() with no
arguments and results in creating an identity matrix.

Change-Id: Id66bb0e1435c015c25aaa717ca98f3aa5d539530
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>