qgltf: Fix not putting textures into the qrc when no copying occurs
Now that the qmake rule uses "." as the default output directory for the
generated files, external assets like textures are often not copied as they
are in the same place already. However, we still need to list everything in
the generated .qrc file.

This should fix not having textures in many cases with the simplified prf.

Change-Id: I2b593adf9ac9dddf32bf1e411bd9443025727f8f
Reviewed-by: Mika Salmela <mika.salmela@theqtcompany.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>