Allow to link on systems with another Qt installed in /usr/lib
Without this commit, the link line for qgltf tries to search in /usr/lib
before the qtbase/lib dir of the Qt currently being built. In doing so
it found the system installed QtCore which is not what we want. In my
case it failed because of missing symbols as I'm using a namespaced
build but in other cases it may lead to more subtle errors.

Using PKGCONFIG_PRIVATE puts the /usr/lib search path from pkgconfig at
the end rather than the beginning of the search path list.

Change-Id: I0f715564c3255100a42d6e9c12b1d10168cd05bb
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>