Add a qgltf example.
This is like gltf but uses the original Collada asset and generates
a glTF model at build time. Now that the qgltf rule is in qtbase we
can freely use it in any examples.

This does not have the limitations the gltf example has: this will
function both with core profiles and ES.

Change-Id: I9f04cf8b77337d397e1aecd1ffb22b7ce7e15ef7
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>