Added support for partial OBJ mesh loading
Added a subMesh string property to the QMesh class.
This is passed to the ObjLoader and, if not empty,
it will only load the sub meshes at match.

The string is assumed to be a regular expression against
which sub mesh names are matched.
If the string is not a valid regex then a valid one
is constructed.

Change-Id: Iecc39af8a41429bb924fb1e3e47d8cbdb9ff1086
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>