Track whether the vao has had attributes specified or not
In some cases the VAO was being created but there wasn't any input
attributes specified on the shader at that time which skipped the
specification of the attributes in updateBuffersAndAttributes(). We now
track whether the VAO has been specified as part of the VAO object
itself.

This way, the first time the VAO is actually used to draw something we
can ensure the attributes have been specified.

Also skip the draw calls if the primitive count is zero.

Change-Id: Iad0a1cff4677ab33369057a6000f0f4a18dc5c58
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>