Add AbstractRenderer interface
The goal is to make the Renderer a subclass of this interface. That way the
QRenderAspect can abstract all the rendering part, which will allow us to not
be tied down to the rendering technology.

We can imagine in a later step that we could add support for DirectX/Vulkan or
any other rendering API by just creating a new IRenderer implementation.

Change-Id: Ie3bd59f5d06c579da81b6e3553f26f1da765101f
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>