Start refactoring the input device plugin
The plugin will now create a QInputDeviceIntegration object which can
act as a "mini input aspect" by setting things up and creating jobs
each frame.

This will also (soon) register the frontend QAbstractPhysicalDevice
subclass with a factory on the QInputAspect to allow C++ users to create
the frontend device type.

Change-Id: I709febfe12b5fae7b6157d1192fb8720362c7622
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>