Add light support to GL2/ES2 default shader
However, take only the first light into account in order to keep pure
ES2 devices working. We will need to figure something out later.

NB this leads to a funny but totally correct effect in examples with
multiple animated lights since the light that is taken into account is
the one closest to the center of the lit entity.

Change-Id: If2ad7c0bd5db177e48d763e1e1c86e65add45c8d
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>