Indicate light type
Having a dedicated type field is required not just to make the shaders
cleaner but also to avoid incorrect results when lights get moved around in
the scene. Currently the shaders rely on the direction uniform to distinguish
directional lights from the others, but this cannot work when the elements
in the lights uniform array change, potentially leaving type-specific members
like direction set when a point light takes the same index.

Change-Id: I170e4b471c8cd9b4a23eca49690239c01944b3e6
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>