Add sensitivity property to QMouseController
The sensitivity is defined as the recipricol of the number of pixels
the mouse is required to move to obtain the extreme axis value (-1 or
+1). That is if you set it to 0.1 you need to move the mouse 10 pixels
to get an axis value of magnitude 1. If it is 0.01, 100 pixels gives
an axis value of magnitude 1.

The user is responsible for applying any additional transformations in
a suitable handler.

Change-Id: I51cd28a17d4fc0b8461ac3e49efb26b2c66081b3
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>