Separate GLSL 1.00 and newer shader snippets for lights
Many but not all OpenGL ES 2.0 implementations support dynamic indexing.
For these the normal for loop is fine. For simple GLSL 1.00 compilers
without dynamic indexing support (e.g. RPi) a separate, limited version
is provided that supports a lower number of lights and uses copy-paste
instead of the loop.

Right now there is no sane way to choose which snippet to use as that
requires the OpenGL context at minimum (and even then we could only rely
on vendor/renderer checks). Therefore a QT3D_GLSL100_WORKAROUND
environment variable is used for now. This is temporary until
we figure out something more sophisticated.

The default GL2/ES2 lighting shader is now equivalent to the GL3
version, supporting up to 8 lights.

Change-Id: I0244c920a1f2c2516b130991996afca68207a7db
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>