Fix Rendering to Scene3D on Raspberry Pi (maybe other ES2 devices)
When using an external FBO like in the case of Scene3D, we try to get
the query the size of the FBO's color attachment so that we can call
glViewport in the case that there are multiple viewports.  For ES2 this
is problematic because it is not possible to query the size of a
GL_TEXTURE attachment (which the attachment will normally be if there is
no multi-sampling).  When the query is not possible we simply return
without calling glViewport. On ES2 it is possible to query the size of a
RenderBuffer and this has been implemented for this code path.

As an oddity in the VideoCore OpenGL drivers for the Raspberry Pi, when
we query the type of the color attachment for the FBO, it returns the
type as RenderBuffer even though we created it as a GL_TEXTURE (likely
because they are the same in the driver).  This creates an issue though
because when we query the size of the RenderBuffer it returns a height
and width of 0, which we then use to call glViewport, leading to nothing
being rendered into the FBO.

This commit checks the validity of the size before making the call to
glViewport.

Change-Id: I0c3235afd7bcc7fe70ca0dda0379b1ceb2cdab38
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>