QTransform: Add convenience properties for Euler angle rotations
Since refactoring the QTransform API to enable matrix decomposition, it
has become more inconvenient to apply rotation transforms through the
QML API.  This is due to the need to set QQuaternion values for the
rotation property, which makes using the Qt Quick animation framework
difficult.

To remedy this, there are now some additional properties added to
Qt3DCore::QTransform which enable direct usage of Euler angles instead
of just QQuaternion.  Internally the rotation property is still used by
translating the Euler angle transforms to a QQuaternion.

Change-Id: Iad48542a7b125f7615f7a69516fb10a973d8395b
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>